Move WalkComponent to components
