+++ engine/core/src/main/java/es/eucm/ead/engine/components/WalkComponent.java
@@ -34,7 +34,7 @@
  *      You should have received a copy of the GNU Lesser General Public License
  *      along with eAdventure.  If not, see <http://www.gnu.org/licenses/>.
  */
-package es.eucm.ead.engine;
+package es.eucm.ead.engine.components;
 
 import ashley.core.Component;
 import com.badlogic.gdx.math.Vector2;
